requirements.txt
# 🏥 AI Healthcare Assistant - Dependencies

# Streamlit and Web Framework
streamlit==1.28.1
streamlit-option-menu==0.3.6
streamlit-webrtc==0.47.1

# AI and Machine Learning
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
transformers==4.33.2
huggingface-hub==0.16.4
opencv-python==4.8.1.78
scikit-learn==1.3.0
scikit-image==0.21.0

# Medical Image Processing
pydicom==2.4.3
SimpleITK==2.3.1
nibabel==5.1.0
pillow==10.0.1

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
speech-recognition==3.10.0
pyaudio==0.2.11

# 3D Visualization
plotly==5.17.0
vtk==9.3.0
trimesh==3.23.5
open3d==0.17.0

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Data Processing
pandas==2.1.1
numpy==1.25.2
matplotlib==3.7.2
seaborn==0.12.2

# API and Networking
requests==2.31.0
fastapi==0.103.2
uvicorn==0.23.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
pathlib==1.0.1

# Database
sqlite3

# Additional AI Tools
openai==0.28.1
langchain==0.0.292
chromadb==0.4.13

# Deployment
gunicorn==21.2.0
docker==6.1.3